Merged 25809:26050 from trunk


svn path=/branches/soc-2008-text/; revision=26051

+++ plug-ins/pygimp/pygimp-pdb.c
@@ -820,7 +820,7 @@ pf_call(PyGimpPDBFunction *self, PyObject *args, PyObject *kwargs)
 #endif
 
     if (kwargs) {
-        int len, pos;
+        Py_ssize_t len, pos;
         PyObject *key, *val;
 
         len = PyDict_Size(kwargs);
